requirements.txt
flask==3.0.3
pandas==2.2.3
requests==2.32.3
pillow==11.0.0
python-dotenv==1.0.1